hot fix

+++ src/lib/cartesio_server.cpp
@@ -55,7 +55,7 @@ CartesIOServer::CartesIOServer(const ros::NodeHandle& node_handle, const ros::No
   for (auto& group_name : group_names_) {
     std::string action_name = "cartesian/" + group_name + "/reach";
     auto client = std::make_shared<reachPoseActionClient>(action_name);
-    if (!client->waitForServer(ros::Duration(double(timeout)))) {
+    if (!client->waitForServer(ros::Duration(int(timeout)))) {
       throw std::runtime_error("RoPort: Action server " + action_name + " unavailable");
     }
     control_clients_.push_back(client);
